+++ src/main/java/com/boleg/crud_pro/entity/Role.java
@@ -12,7 +12,7 @@ public class Role implements GrantedAuthority {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "role")
     private String role;
@@ -20,16 +20,16 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(int id, String role) {
+    public Role(Long id, String role) {
         this.id = id;
         this.role = role;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
